requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1.78
pillow==10.0.0
joblib==1.3.2
pyyaml==6.0.1
jupyter==1.0.0
notebook==6.5.4
Flask==2.3.3
Werkzeug==2.3.7